backend/requirements.txt
# Scraping
scrapy==2.11.2
playwright==1.44.0
scrapy-playwright==0.0.34

# PDF parsing
pymupdf==1.24.11
pdfplumber==0.11.1

# Embeddings + Vector DB
openai==1.35.0
pinecone==4.1.0

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
tiktoken==0.7.0